Move sync logic into separate classes
