+++ xrateskit/src/main/java/io/horizontalsystems/xrateskit/CoinExternalIdsSyncer.kt
@@ -1,17 +1,21 @@
 package io.horizontalsystems.xrateskit
 
 import io.horizontalsystems.xrateskit.entities.ProviderCoinsResource
+import io.horizontalsystems.xrateskit.entities.ResourceInfo
 import io.horizontalsystems.xrateskit.entities.ResourceType
 import io.horizontalsystems.xrateskit.storage.Storage
 
-class ProviderCoinsResourceManager(
+class CoinExternalIdsSyncer(
     private val providerCoinsResourceProvider: ProviderCoinsResourceProvider,
     private val storage: Storage
 ) {
-    fun getNewData() : ProviderCoinsResource? {
+    fun syncData() {
         val resourceInfo = storage.getResourceInfo(ResourceType.PROVIDER_COINS)
 
-        return providerCoinsResourceProvider.getDataNewerThan(resourceInfo?.version)
+        providerCoinsResourceProvider.getDataNewerThan(resourceInfo?.version)?.let {
+            storage.saveProviderCoins(it.providerCoins)
+            storage.saveResourceInfo(ResourceInfo(ResourceType.PROVIDER_COINS, it.version))
+        }
     }
 
 }
