--- xrateskit/src/main/java/io/horizontalsystems/xrateskit/coins/provider/CoinInfoResourceManager.kt
@@ -1,16 +1,0 @@
-package io.horizontalsystems.xrateskit.coins.provider
-
-import io.horizontalsystems.xrateskit.entities.CoinInfoResource
-import io.horizontalsystems.xrateskit.entities.ResourceType
-import io.horizontalsystems.xrateskit.storage.Storage
-
-class CoinInfoResourceManager(
-    private val coinInfoResourceProvider: CoinInfoResourceProvider,
-    private val storage: Storage
-) {
-    fun getNewData(): CoinInfoResource? {
-        val resourceInfo = storage.getResourceInfo(ResourceType.COIN_INFO)
-
-        return coinInfoResourceProvider.getDataNewerThan(resourceInfo?.version)
-    }
-}
