+++ xrateskit/src/main/java/io/horizontalsystems/xrateskit/coins/provider/CoinInfoSyncer.kt
@@ -1,0 +1,34 @@
+package io.horizontalsystems.xrateskit.coins.provider
+
+import io.horizontalsystems.xrateskit.entities.ResourceInfo
+import io.horizontalsystems.xrateskit.entities.ResourceType
+import io.horizontalsystems.xrateskit.storage.Storage
+
+class CoinInfoSyncer(
+    private val coinInfoResourceProvider: CoinInfoResourceProvider,
+    private val storage: Storage
+) {
+    fun syncData() {
+        val resourceInfo = storage.getResourceInfo(ResourceType.COIN_INFO)
+
+        coinInfoResourceProvider.getDataNewerThan(resourceInfo?.version)?.let { coinsResponse ->
+            storage.deleteAllCoinCategories()
+            storage.deleteAllCoinLinks()
+            storage.deleteAllCoinsCategories()
+            storage.deleteAllCoinFunds()
+            storage.deleteAllCoinsFunds()
+            storage.deleteAllCoinFundCategories()
+            storage.deleteAllExchangeInfo()
+
+            storage.saveCoinInfos(coinsResponse.coinInfos)
+            storage.saveCoinCategories(coinsResponse.coinsCategories)
+            storage.saveCoinCategory(coinsResponse.categories)
+            storage.saveCoinFund(coinsResponse.funds)
+            storage.saveCoinFunds(coinsResponse.coinFunds)
+            storage.saveCoinFundCategory(coinsResponse.fundCategories)
+            storage.saveCoinLinks(coinsResponse.links)
+            storage.saveExchangeInfo(coinsResponse.exchangeInfos)
+            storage.saveResourceInfo(ResourceInfo(ResourceType.COIN_INFO, coinsResponse.version))
+        }
+    }
+}
